+++ R/DA-ancom.R
@@ -104,6 +104,8 @@ run_ancom <- function(ps,
     test = c("aov", "wilcox.test", "kruskal.test"),
     ...) {
     stopifnot(inherits(ps, "phyloseq"))
+    ps <- check_rank_names(ps) %>% 
+        check_taxa_rank( taxa_rank)
     test <- match.arg(test, c("aov", "wilcox.test", "kruskal.test"))
 
     # check whether group is valid, write a function
@@ -132,19 +134,7 @@ run_ancom <- function(ps,
     # normalize the data
     norm_para <- c(norm_para, method = norm, object = list(ps))
     ps_normed <- do.call(normalize, norm_para)
-
-    # summarize data
-    # create a function, extract_summarize
-    # check taxa_rank
-    check_taxa_rank(ps, taxa_rank)
-    if (taxa_rank == "all") {
-        ps_summarized <- summarize_taxa(ps_normed)
-    } else if (taxa_rank == "none") {
-        ps_summarized <- extract_rank(ps_normed, taxa_rank)
-    } else {
-        ps_summarized <- aggregate_taxa(ps_normed, taxa_rank) %>%
-            extract_rank(taxa_rank)
-    }
+    ps_summarized <- pre_ps_taxa_rank(ps_normed, taxa_rank)
 
     feature_table <- abundances(ps_summarized, norm = TRUE)
     meta_data <- data.frame(sample_data(ps_summarized))