+++ tests/testthat/test-ancom.R
@@ -43,9 +43,6 @@ test_that("ancom result", {
         taxa_rank = "Class"
     )
     curr_marker <- marker_table(ancom_res)
-    expect_known_output(
-        print(head(curr_marker), digits = 5),
-        test_path("out/test-ancom_marker.txt"),
-        print = TRUE
-    )
+    
+    expect_snapshot(print(head(curr_marker), digits = 5))
 })